Use ES6 imports for all JS
