+++ index.html
@@ -19,16 +19,8 @@
     <link rel="stylesheet" href="web_src/css/L.Control.MousePosition.css" />
     <link rel="stylesheet" href="web_src/css/Control.FullScreen.css" />
     <link rel="stylesheet" href="web_src/css/leaflet-search.src.css" />
-
     <link rel="stylesheet" href="web_src/css/sidebar.css" />
     <link rel="stylesheet" href="web_src/css/main.css" />
-
-    <script src="js/lib/leaflet.min.js"></script>
-    <script src="js/lib/leaflet.toolbar.min.js"></script>
-    <script src="js/lib/L.TileLayer.Canvas.js"></script>
-    <script src="js/lib/Control.FullScreen.js"></script>
-    <script src="js/lib/leaflet-search.src.min.js"></script>
-    <script src="js/lib/L.Control.MousePosition.js"></script>
   </head>
   <body>
     <div id="sidebar_container"></div>
