+++ package.json
@@ -13,6 +13,7 @@
   "dependencies": {
     "@fortawesome/react-fontawesome": "^3.3.1",
     "classnames": "^2.5.1",
+    "leaflet": "^1.9.4",
     "react": "^19.2.4",
     "react-dom": "^19.2.4"
   },
