+++ web_src/lib/L.Control.MousePosition.js
